feat(M16 C2): D7 hot-path mirror dispatch + D33 counter triplet
Lands the real mirror verb on the hot path:

- src/dataplane/action_dispatch.h: new stage_mirror/mirror_drain
  helpers modeled on the D16 stage_redirect/redirect_drain pattern;
  `case kMirror:` in apply_action now clones via rte_pktmbuf_copy,
  forwards the original via tx_one(ctx->tx_port_id), stages the
  clone for end-of-burst drain. rte_pktmbuf_copy==nullptr path
  bumps mirror_clone_failed_per_port and keeps the original
  forwarding.
- src/dataplane/worker.h: kMirrorBurstMax=32 + MirrorBuf struct;
  mirror_tx[RTE_MAX_ETHPORTS] + mirror_{sent,clone_failed,dropped}
  _per_port[RTE_MAX_ETHPORTS] on WorkerCtx (768 B per worker).
- src/dataplane/worker.cpp: mirror_drain(ctx) runs adjacent to
  redirect_drain at burst end.
- D33 six-fold lockstep: counter_names.h grows 39 -> 41 with
  kMirrorSentTotal + kMirrorCloneFailedTotal (the pre-existing
  placeholder kMirrorDroppedTotal reused); snapshot.h +
  snapshot.cpp aggregate the three per-port arrays into parallel
  vectors, snapshot_metric_names() emits all three base names;
  main.cpp wires LcoreCounterView + BodyFn /metrics emit; design.md
  §10.3 grows the three rows; scripts/check-counter-consistency.sh
  drops the transient `pktgate_mirror_` allowlist entry (exit 0
  verified).

D1 / M11 C1.5 hot path: every bump goes through relaxed_bump /
relaxed_bump_bucket / relaxed_add. Objdump of worker.cpp.o shows
zero `lock` prefixes in the mirror dispatch / stage / drain code;
the only 5 lock instructions in the TU are 4 pre-existing
`lock decw` inside rte_pktmbuf_free (DPDK refcnt helper) + 1
`lock addl $0,(%rsp)` inside worker_main at the RCU thread-online
registration (worker entry only, not per-packet). Paste:

  $ objdump -d build/dev-release/.../worker.cpp.o | awk '...'
  rte_pktmbuf_free:       f8: lock decw (%rax)
  rte_pktmbuf_free:      15b: lock decw (%rsi)
  rte_pktmbuf_free:      347: lock decw (%rsi)
  rte_pktmbuf_free:      356: lock decw (%rsi)
  worker_main:           419: lock addl $0x0,-0x80(%rsp)  # RCU reg

Full 5-preset matrix (M15 C2 amendment 2026-04-20), sudo ctest -L
"smoke|unit|functional|integration|chaos" -j 1 --output-on-failure
on ssh dpdk:

  === dev-debug ===
  100% tests passed, 0 tests failed out of 53
  Total Test time (real) = 286.77 sec

  === dev-release ===
  100% tests passed, 0 tests failed out of 53
  Total Test time (real) = 287.36 sec

  === dev-asan ===
  100% tests passed, 0 tests failed out of 53
  Total Test time (real) = 300.44 sec

  === dev-ubsan ===
  100% tests passed, 0 tests failed out of 53
  Total Test time (real) = 289.06 sec

  === dev-tsan ===
  100% tests passed, 0 tests failed out of 53
  Total Test time (real) = 423.72 sec

D33 gate:

  $ bash scripts/check-counter-consistency.sh
  pass1: canonical set = 41 metric(s) from §10.3
  pass2: ok — 3 compiler counter families verified, 33 deferred
  pass3: ok — every prose pktgate_* token is in §10.3 (or allowlisted)

C7.27 manifest test: 41 canonical entries; RED at 75ef346 required
mirror_sent / mirror_clone_failed; both land in kAllCounterNames
and canonical_manifest in lockstep. U16.6..U16.11 + F16.1 in
unit_test_eal_unit all green under all five presets.

Refs D7, D33, D41, D1, M11 C1.5.
Anti-pivot: RED at 75ef346, GREEN here — no --allow-empty path.

+++ tests/unit/test_eal_unit.cpp
@@ -9795,7 +9795,7 @@ TEST_F(MirrorDispatchTest, U16_6_StageMirrorBuffersClone) {
 
   // Drain for cleanup — spy "sends" the clone; free staged pointer.
   dataplane::mirror_drain(&ctx);
-  for (std::uint16_t i = 0; i + 1 < spy.seen_mbufs.size(); ++i) {
+  for (std::size_t i = 0; i + 1 < spy.seen_mbufs.size(); ++i) {
     // pkts[0..n-1] in spy.seen_mbufs past index 0 are the clones; the
     // spy holds but does not own. Drain already removed them from
     // mirror_tx; free them here.
